Create & Delete product api

+++ ecommerce-backend/src/main/java/com/abdullah/eCommerce/dtos/OrderItemDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,5 +15,5 @@ public class OrderItemDto {
     private Long productId;
     private String productTitle;
     private int quantity;
-    private int pricePerUnit;
+    private BigDecimal pricePerUnit;
 }
